move & rename file, add preconnect method
